+++ MotivateDev/app/src/main/java/com/productions/itea/motivatedev/UserScore.java
@@ -5,19 +5,19 @@ import android.net.Uri;
 
 public class UserScore {
     public String username;
-    public Long score;
+    public Integer score;
     public String photoUrl;
 
     public UserScore() {
         // Default constructor required for calls to DataSnapshot.getValue(User.class)
     }
 
-    UserScore (String name, Long score, Uri photoUrl) {
+    UserScore (String name, Integer score, Uri photoUrl) {
         this.username = name;
         this.score = score;
         this.photoUrl = photoUrl != null ? photoUrl.toString() : "";
     }
-    UserScore (String name, Long score, String photoUrl) {
+    UserScore (String name, Integer score, String photoUrl) {
         this.username = name;
         this.score = score;
         this.photoUrl = photoUrl;
